Modularise a few more things
