+++ src/components/Molecules/ArticleTeaser/ArticleTeaser.js
@@ -10,11 +10,7 @@ import {
 import altCtaUnderline from '../../../theme/shared/assets/alt_cta_underline.svg';
 import { ArrowIconWrapper, ArrowIconInner } from '../shared/ctaArrow/CtaArrowCircle.style';
 import ArrowIcon from '../shared/ctaArrow/ArrowIcon';
-import CtaTextUnderline from '../shared/CtaUnderline/CtaUnderline.style';
-
-import {
-  CtaIconWrapper, CtaText
-} from '../HeroBanner/HeroBanner.style';
+import { CtaTextUnderline, CtaIconWrapper, CtaText } from '../shared/ctaText/ctaText.style';
 
 const handleCampaignLogo = category => {
   switch (category) {
