--- src/components/Molecules/shared/CtaUnderline/CtaUnderline.style.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const CtaTextUnderline = styled.img`
-  height: 4px;
-  width: 100%;
-  position: absolute;
-  left: 0;
-  bottom: -5px;
-  transition: opacity 0.15s 0.1s;
-  opacity: 0;
-`;
-
-export default CtaTextUnderline;
